web_pytest_2020/requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
PyExecJS==1.5.1
Pygments==2.9.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==4.5.0
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.11.0
pytest-repeat==0.7.0
pytest-rerunfailures==9.0
pytest-xdist==1.23.2
pytz==2021.1
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.9.1
selenium==3.141.0
sentry-sdk==0.13.5
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.22
wcwidth==0.2.5
xeger==0.3.5
yarg==0.1.9
zipp==3.4.0
